+++ lib/queen_validators.dart
@@ -1,7 +1,17 @@
 library queen_validators;
 
+/// `qValidattor` function
 export 'package:queen_validators/src/validator.dart';
+
+/// `abstract` `class` to impelemnt your own DTO || rule
 export 'package:queen_validators/src/imp/dto.dart';
+export 'package:queen_validators/src/imp/validation.dart';
+
+/// export aviliable rule
 export 'package:queen_validators/src/rules/index.dart';
+
+/// export aviliable extensions methods to validate variables
 export 'package:queen_validators/src/extensions/index.dart';
+
+/// only supported exception which will be trown in case of invalid DTO
 export 'package:queen_validators/src/exception.dart';