+++ lib/src/rules/is_in.dart
@@ -2,7 +2,7 @@ import 'package:queen_validators/src/imp/validation.dart';
 import 'package:queen_validators/src/extensions/string.dart';
 
 /// checks if the input `is in` provided List;
-class IsIn extends QueenValidationRule<String> {
+class IsIn extends QueenValidationRule {
   final List<String> list;
   IsIn(this.list, {String msg})
       : assert(list != null),