+++ lib/src/rules/is_not_in.dart
@@ -2,7 +2,7 @@ import 'package:queen_validators/src/imp/validation.dart';
 import 'package:queen_validators/src/extensions/string.dart';
 
 /// checks if the input `is NOT in` provided List;
-class IsNotIn extends QueenValidationRule<String> {
+class IsNotIn extends QueenValidationRule {
   final List<String> list;
   IsNotIn(this.list, {String msg})
       : assert(list != null),