+++ lib/src/rules/is_optional.dart
@@ -3,7 +3,7 @@ import 'package:queen_validators/src/imp/validation.dart';
 /// allows you to skip the errors if the input is null
 /// * if the input is null and fails any role
 /// * with instance of this class the qValidator will ignore the faild rule
-class IsOptional extends QueenValidationRule<String> {
+class IsOptional extends QueenValidationRule {
   IsOptional() : super(null);
 
   @override