+++ lib/src/rules/is_required.dart
@@ -2,7 +2,7 @@ import 'package:queen_validators/queen_validators.dart';
 import 'package:queen_validators/src/imp/validation.dart';
 
 /// checks if the input is not empty string
-class IsRequired extends QueenValidationRule<String> {
+class IsRequired extends QueenValidationRule {
   IsRequired({String msg}) : super(msg);
 
   @override