+++ lib/src/rules/match.dart
@@ -2,7 +2,7 @@ import 'package:queen_validators/src/imp/validation.dart';
 import 'package:queen_validators/src/extensions/string.dart';
 
 /// checks if the input is match other string
-class Match extends QueenValidationRule<String> {
+class Match extends QueenValidationRule {
   final String stringToMatchWith;
   final bool ignoreCase;
   Match(this.stringToMatchWith, {String msg, this.ignoreCase})