+++ lib/src/rules/max_length.dart
@@ -1,7 +1,7 @@
 import 'package:queen_validators/src/imp/validation.dart';
 
 /// checks if the input charachtrs length is smaller than the min feild
-class MaxLength extends QueenValidationRule<String> {
+class MaxLength extends QueenValidationRule {
   final int max;
   final bool trim;
   MaxLength(this.max, {String msg, this.trim = true}) : super(msg);
@@ -10,6 +10,5 @@ class MaxLength extends QueenValidationRule<String> {
   String get errorMsg => 'max length is $max';
 
   @override
-  bool validate(String val) =>
-      trim ? val.trim().length < max : val.length < max;
+  bool validate(String val) => trim ? val.trim().length <= max : val.length <= max;
 }