+++ lib/src/rules/min_length.dart
@@ -2,7 +2,7 @@ import 'package:queen_validators/src/imp/validation.dart';
 
 /// checks if the input charachtrs length is bigger than the min feild
 
-class MinLength extends QueenValidationRule<String> {
+class MinLength extends QueenValidationRule {
   final int min;
   final bool trim;
 
@@ -12,6 +12,5 @@ class MinLength extends QueenValidationRule<String> {
   String get errorMsg => 'min length is $min';
 
   @override
-  bool validate(String val) =>
-      trim ? val.trim().length > min : val.length > min;
+  bool validate(String val) => trim ? val.trim().length >= min : val.length >= min;
 }