+++ lib/src/rules/regx.dart
@@ -1,7 +1,7 @@
 import 'package:queen_validators/src/imp/validation.dart';
 
 /// allow using a regular expreesion as validation rule
-class RegExpRule extends QueenValidationRule<String> {
+class RegExpRule extends QueenValidationRule {
   final RegExp regX;
   RegExpRule(this.regX, {String msg}) : super(msg);